_scripts/requirements.txt
openai
feedparser
youtube-transcript-api
yt-dlp
requests
mechanicalsoup
python-dotenv
pypdf
beautifulsoup4
